Wrong type in method. Int should be Double
``def getComputedTextLength(): Int = js.native``
should be 
def getComputedTextLength(): Double = js.native

+++ src/main/scala/org/scalajs/dom/raw/Svg.scala
@@ -1895,7 +1895,7 @@ abstract class SVGTextContentElement
 
   def getExtentOfChar(charnum: Int): SVGRect = js.native
 
-  def getComputedTextLength(): Int = js.native
+  def getComputedTextLength(): Double = js.native
 
   def getSubStringLength(charnum: Int, nchars: Int): Int = js.native
